interface를 abstract class로 변경해 공통구현부분을 추가
