+++ app/src/main/java/net/jspiner/epub_viewer/ui/search/finder/ScrollPageFinder.kt
@@ -12,11 +12,8 @@ import io.reactivex.subjects.SingleSubject
 import net.jspiner.epub_viewer.dto.Epub
 import net.jspiner.epub_viewer.dto.PageInfo
 import net.jspiner.epubstream.dto.ItemRef
-import java.io.BufferedReader
-import java.io.File
-import java.io.FileReader
 
-class ScrollPageFinder(val context: Context, val epub: Epub, val pageInfo: PageInfo) : PageFinder {
+class ScrollPageFinder(context: Context, epub: Epub, pageInfo: PageInfo) : PageFinder(context, epub, pageInfo) {
 
     private val deviceWidth: Int by lazy { context.resources.displayMetrics.widthPixels }
     private val deviceHeight: Int by lazy { context.resources.displayMetrics.heightPixels }
@@ -52,20 +49,6 @@ class ScrollPageFinder(val context: Context, val epub: Epub, val pageInfo: PageI
             .subscribeOn(Schedulers.io())
     }
 
-    private fun readFile(file: File): String {
-        return BufferedReader(FileReader(file)).use { br ->
-            val sb = StringBuilder()
-            var line = br.readLine()
-
-            while (line != null) {
-                sb.append(line)
-                sb.append(System.lineSeparator())
-                line = br.readLine()
-            }
-            sb.toString()
-        }
-    }
-
     private fun setUpWebView(webView: WebView) {
         webView.isVerticalScrollBarEnabled = true
         webView.settings.apply {
@@ -92,16 +75,5 @@ class ScrollPageFinder(val context: Context, val epub: Epub, val pageInfo: PageI
             val webViewHeight = height * context.resources.displayMetrics.density
             subject.onSuccess(webViewHeight.toLong())
         }
-    }
-
-    private fun toManifestItem(itemRef: ItemRef): File {
-        val manifestItemList = epub.opf.manifest.items
-
-        for (item in manifestItemList) {
-            if (item.id == itemRef.idRef) {
-                return epub.extractedDirectory.resolve(item.href)
-            }
-        }
-        throw RuntimeException("해당 itemRef 를 manifest 에서 찾을 수 없음 id : ${itemRef.idRef}")
     }
 }
